requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
python-dotenv>=1.0.0
google-generativeai>=0.5.0
streamlit>=1.35.0
requests>=2.31.0
python-multipart>=0.0.6
pydantic>=2.0.0
